Display autoMLmodel output in HTML format using Rmarkdown

This function will generate R markdown report for DriveML model object

autoMLReport(mlobject, mldata = NULL, op_file = NULL, op_dir = NULL)

Arguments

mlobject
[autoMLmodel Object | Required] autoMLmodel function output

mldata
[autoDataprep Object | Optional] autoDataprep function output

op_file
[character | Required] output file name (.html)

op_dir
[character | Optional] output path. Default path is the current working directory

Value

HTML R Markdown output

Details

Using this function we can easily present the model outcome in standard HTML format without writing Rmarkdown scripts

Examples

## Creating HTML report

 autoMLReport(heart.model, mldata = NULL, op_file = "sample.html", op_dir = tempdir())

#> 
#> 
#> processing file: report_tmp.Rmd

#>   |                                                                         |                                                                 |   0%  |                                                                         |.                                                                |   2%
#>    inline R code fragments
#> 
#>   |                                                                         |...                                                              |   4%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                         |....                                                             |   6%
#>   ordinary text without R code
#> 
#>   |                                                                         |.....                                                            |   8%
#> label: dl_1 (with options) 
#> List of 3
#>  $ warning: logi FALSE
#>  $ eval   : symbol T
#>  $ include: symbol F
#> 
#>   |                                                                         |.......                                                          |  10%
#>   ordinary text without R code
#> 
#>   |                                                                         |........                                                         |  12%
#> label: dl_11 (with options) 
#> List of 4
#>  $ warning: logi FALSE
#>  $ eval   : symbol T
#>  $ render : symbol t1
#>  $ echo   : symbol F
#> 
#>   |                                                                         |.........                                                        |  14%
#>   ordinary text without R code
#> 
#>   |                                                                         |...........                                                      |  16%
#> label: dl_12 (with options) 
#> List of 4
#>  $ warning: logi FALSE
#>  $ eval   : symbol T
#>  $ render : symbol t2
#>  $ echo   : symbol F
#> 
#>   |                                                                         |............                                                     |  18%
#>   ordinary text without R code
#> 
#>   |                                                                         |.............                                                    |  20%
#> label: dl_13 (with options) 
#> List of 4
#>  $ warning: logi FALSE
#>  $ eval   : symbol T
#>  $ render : symbol t3
#>  $ echo   : symbol F
#> 
#>   |                                                                         |...............                                                  |  22%
#>   ordinary text without R code
#> 
#>   |                                                                         |................                                                 |  24%
#> label: dl_14 (with options) 
#> List of 4
#>  $ warning: logi FALSE
#>  $ eval   : symbol T
#>  $ render : symbol mdata
#>  $ echo   : symbol F
#> 
#>   |                                                                         |.................                                                |  27%
#>   ordinary text without R code
#> 
#>   |                                                                         |...................                                              |  29%
#> label: dl_15 (with options) 
#> List of 4
#>  $ warning: logi FALSE
#>  $ eval   : symbol T
#>  $ render : symbol drmodel
#>  $ echo   : symbol F
#> 
#>   |                                                                         |....................                                             |  31%
#>   ordinary text without R code
#> 
#>   |                                                                         |.....................                                            |  33%
#> label: d2 (with options) 
#> List of 4
#>  $ warning: logi FALSE
#>  $ eval   : symbol T
#>  $ render : symbol result
#>  $ echo   : symbol F
#> 
#>   |                                                                         |.......................                                          |  35%
#>   ordinary text without R code
#> 
#>   |                                                                         |........................                                         |  37%
#> label: std1 (with options) 
#> List of 7
#>  $ fig.show : chr "hold"
#>  $ out.width: chr "50%"
#>  $ fig.align: chr "default"
#>  $ warning  : logi FALSE
#>  $ eval     : symbol pl_glmnet
#>  $ render   : symbol modelobject
#>  $ echo     : symbol F
#> 

#>   |                                                                         |.........................                                        |  39%
#>   ordinary text without R code
#> 
#>   |                                                                         |...........................                                      |  41%
#> label: std2 (with options) 
#> List of 7
#>  $ fig.show : chr "hold"
#>  $ out.width: chr "50%"
#>  $ fig.align: chr "default"
#>  $ warning  : logi FALSE
#>  $ eval     : symbol pl_logreg
#>  $ render   : symbol modelobject
#>  $ echo     : symbol F
#> 

#>   |                                                                         |............................                                     |  43%
#>   ordinary text without R code
#> 
#>   |                                                                         |.............................                                    |  45%
#> label: std3 (with options) 
#> List of 7
#>  $ fig.show : chr "hold"
#>  $ out.width: chr "50%"
#>  $ fig.align: chr "default"
#>  $ warning  : logi FALSE
#>  $ eval     : symbol pl_randomForest
#>  $ render   : symbol modelobject
#>  $ echo     : symbol F
#> 

#>   |                                                                         |...............................                                  |  47%
#>   ordinary text without R code
#> 
#>   |                                                                         |................................                                 |  49%
#> label: std4 (with options) 
#> List of 7
#>  $ fig.show : chr "hold"
#>  $ out.width: chr "50%"
#>  $ fig.align: chr "default"
#>  $ warning  : logi FALSE
#>  $ eval     : symbol pl_ranger
#>  $ render   : symbol modelobject
#>  $ echo     : symbol F
#> 

#>   |                                                                         |.................................                                |  51%
#>   ordinary text without R code
#> 
#>   |                                                                         |..................................                               |  53%
#> label: std5 (with options) 
#> List of 7
#>  $ fig.show : chr "hold"
#>  $ out.width: chr "50%"
#>  $ fig.align: chr "default"
#>  $ warning  : logi FALSE
#>  $ eval     : symbol pl_xgboost
#>  $ render   : symbol modelobject
#>  $ echo     : symbol F
#> 

#>   |                                                                         |....................................                             |  55%
#>   ordinary text without R code
#> 
#>   |                                                                         |.....................................                            |  57%
#> label: std6 (with options) 
#> List of 7
#>  $ fig.show : chr "hold"
#>  $ out.width: chr "50%"
#>  $ fig.align: chr "default"
#>  $ warning  : logi FALSE
#>  $ eval     : symbol pl_rpart
#>  $ render   : symbol modelobject
#>  $ echo     : symbol F
#> 

#>   |                                                                         |......................................                           |  59%
#>   ordinary text without R code
#> 
#>   |                                                                         |........................................                         |  61%
#> label: vip1 (with options) 
#> List of 8
#>  $ fig.show  : chr "hold"
#>  $ out.width : chr "70%"
#>  $ fig.align : chr "default"
#>  $ warning   : logi FALSE
#>  $ eval      : symbol vi_xgboost
#>  $ render    : symbol modelobject
#>  $ echo      : symbol F
#>  $ fig.height: num 7
#> 

#>   |                                                                         |.........................................                        |  63%
#>   ordinary text without R code
#> 
#>   |                                                                         |..........................................                       |  65%
#> label: vip2 (with options) 
#> List of 8
#>  $ fig.show  : chr "hold"
#>  $ out.width : chr "70%"
#>  $ fig.align : chr "default"
#>  $ warning   : logi FALSE
#>  $ eval      : symbol vi_randomForest
#>  $ render    : symbol modelobject
#>  $ echo      : symbol F
#>  $ fig.height: num 7
#> 

#>   |                                                                         |............................................                     |  67%
#>   ordinary text without R code
#> 
#>   |                                                                         |.............................................                    |  69%
#> label: vip3 (with options) 
#> List of 8
#>  $ fig.show  : chr "hold"
#>  $ out.width : chr "70%"
#>  $ fig.align : chr "default"
#>  $ warning   : logi FALSE
#>  $ eval      : symbol vi_ranger
#>  $ render    : symbol modelobject
#>  $ echo      : symbol F
#>  $ fig.height: num 7
#> 

#>   |                                                                         |..............................................                   |  71%
#>   ordinary text without R code
#> 
#>   |                                                                         |................................................                 |  73%
#> label: vip4 (with options) 
#> List of 8
#>  $ fig.show  : chr "hold"
#>  $ out.width : chr "70%"
#>  $ fig.align : chr "default"
#>  $ warning   : logi FALSE
#>  $ eval      : symbol vi_rpart
#>  $ render    : symbol modelobject
#>  $ echo      : symbol F
#>  $ fig.height: num 7
#> 

#>   |                                                                         |.................................................                |  76%
#>   ordinary text without R code
#> 
#>   |                                                                         |..................................................               |  78%
#> label: vip5 (with options) 
#> List of 8
#>  $ fig.show  : chr "hold"
#>  $ out.width : chr "70%"
#>  $ fig.align : chr "default"
#>  $ warning   : logi FALSE
#>  $ eval      : symbol vi_glmnet
#>  $ render    : symbol modelobject
#>  $ echo      : symbol F
#>  $ fig.height: num 7
#> 

#>   |                                                                         |....................................................             |  80%
#>   ordinary text without R code
#> 
#>   |                                                                         |.....................................................            |  82%
#> label: vip6 (with options) 
#> List of 8
#>  $ fig.show  : chr "hold"
#>  $ out.width : chr "70%"
#>  $ fig.align : chr "default"
#>  $ warning   : logi FALSE
#>  $ eval      : symbol vi_logreg
#>  $ render    : symbol modelobject
#>  $ echo      : symbol F
#>  $ fig.height: num 7
#> 

#>   |                                                                         |......................................................           |  84%
#>   ordinary text without R code
#> 
#>   |                                                                         |........................................................         |  86%
#> label: lift1 (with options) 
#> List of 8
#>  $ fig.show  : chr "hold"
#>  $ out.width : chr "100%"
#>  $ fig.align : chr "default"
#>  $ warning   : logi FALSE
#>  $ eval      : symbol T
#>  $ render    : symbol modelobject
#>  $ echo      : symbol F
#>  $ fig.height: num 4
#> 

#>   |                                                                         |.........................................................        |  88%
#>   ordinary text without R code
#> 
#>   |                                                                         |..........................................................       |  90%
#> label: lift2 (with options) 
#> List of 4
#>  $ warning: logi FALSE
#>  $ eval   : symbol T
#>  $ render : symbol modelobject
#>  $ echo   : symbol F
#> 
#>   |                                                                         |............................................................     |  92%
#>   ordinary text without R code
#> 
#>   |                                                                         |.............................................................    |  94%
#> label: pdppp (with options) 
#> List of 9
#>  $ fig.show  : chr "hold"
#>  $ out.width : chr "50%"
#>  $ fig.align : chr "default"
#>  $ warning   : logi FALSE
#>  $ eval      : symbol T
#>  $ render    : symbol modelobject
#>  $ echo      : symbol F
#>  $ fig.height: num 4
#>  $ results   : chr "hide"
#> 
#>   |                                                                         |..............................................................   |  96%
#>   ordinary text without R code
#> 
#>   |                                                                         |................................................................ |  98%
#> label: dat (with options) 
#> List of 4
#>  $ warning: logi FALSE
#>  $ eval   : symbol T
#>  $ render : symbol modelobject
#>  $ echo   : symbol F
#> 
#>   |                                                                         |.................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: C:/Users/dubrangala/AppData/Local/Temp/RtmpIbf9Xj/report_tmp.knit.md

#> "C:/Program Files/RStudio/bin/pandoc/pandoc" +RTS -K512m -RTS "C:/Users/dubrangala/AppData/Local/Temp/RtmpIbf9Xj/report_tmp.utf8.md" --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output pandocfefcb5a7a.html --email-obfuscation none --self-contained --standalone --section-divs --table-of-contents --toc-depth 6 --template "C:\Users\DUBRAN~1\ONEDRI~1\VMWARE~1\DOCUME~1\R\WIN-LI~1\3.6\RMARKD~1\rmd\h\DEFAUL~1.HTM" --no-highlight --variable highlightjs=1 --variable "theme:cerulean" --include-in-header "C:\Users\DUBRAN~1\AppData\Local\Temp\RtmpIbf9Xj\rmarkdown-strfefc7e187654.html" --mathjax --variable "mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML" --lua-filter "C:/Users/dubrangala/OneDrive - VMware, Inc/VMwareCorp/Documents/R/win-library/3.6/rmarkdown/rmd/lua/pagebreak.lua" --lua-filter "C:/Users/dubrangala/OneDrive - VMware, Inc/VMwareCorp/Documents/R/win-library/3.6/rmarkdown/rmd/lua/latex-div.lua" 

#> 
#> Output created: C:/Users/dubrangala/AppData/Local/Temp/RtmpIbf9Xj/sample.html

#> 
#> 
#> Report is generated at "C:\Users\DUBRAN~1\AppData\Local\Temp\RtmpIbf9Xj/sample.html".